Review changes
- Merged previous tally.rs with storage.rs
- Renamed storage.rs to tally.rs
- Fixed mutable functions and references
